Feat: update dependecy

+++ examples/cinc17/build_datasets.py
@@ -20,7 +20,7 @@ def load_all(data_path):
         ecg_file = os.path.join(data_path, record + ".mat")
         ecg_file = os.path.abspath(ecg_file)
         ecg = load_ecg_mat(ecg_file)
-        num_labels = ecg.shape[0] / STEP
+        num_labels = int(ecg.shape[0] / STEP)
         dataset.append((ecg_file, [label]*num_labels))
     return dataset
